impl/web_app/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file='.\requirements.txt' '.\requirements.in'
#
aiohappyeyeballs==2.4.6
    # via aiohttp
aiohttp==3.11.12
    # via
    #   -r .\requirements.in
    #   semantic-kernel
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   httpx
    #   openai
    #   starlette
attrs==25.1.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
azure-core==1.32.0
    # via
    #   azure-cosmos
    #   azure-identity
    #   azure-storage-blob
azure-cosmos==4.9.0
    # via -r .\requirements.in
azure-identity==1.20.0
    # via
    #   -r .\requirements.in
    #   semantic-kernel
azure-storage-blob==12.24.1
    # via -r .\requirements.in
beautifulsoup4==4.13.3
    # via -r .\requirements.in
black==25.1.0
    # via -r .\requirements.in
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
chardet==5.2.0
    # via prance
charset-normalizer==3.4.1
    # via
    #   httpie
    #   requests
click==8.1.8
    # via
    #   black
    #   uvicorn
cloudevents==1.11.0
    # via semantic-kernel
colorama==0.4.6
    # via
    #   click
    #   httpie
    #   pytest
    #   tqdm
coverage[toml]==7.6.12
    # via pytest-cov
cryptography==44.0.1
    # via
    #   -r .\requirements.in
    #   azure-identity
    #   azure-storage-blob
    #   msal
    #   pyjwt
defusedxml==0.7.1
    # via
    #   httpie
    #   semantic-kernel
deprecated==1.2.18
    # via
    #   opentelemetry-api
    #   opentelemetry-semantic-conventions
deprecation==2.1.0
    # via cloudevents
distro==1.9.0
    # via openai
docopt==0.6.2
    # via -r .\requirements.in
faker==36.1.1
    # via -r .\requirements.in
fastapi==0.115.8
    # via
    #   -r .\requirements.in
    #   fastapi-msal
fastapi-msal==2.1.6
    # via -r .\requirements.in
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via
    #   httpcore
    #   hypercorn
    #   uvicorn
    #   wsproto
h2==4.2.0
    # via hypercorn
hpack==4.1.0
    # via h2
httpcore==1.0.7
    # via httpx
httpie==3.2.4
    # via -r .\requirements.in
httpx==0.28.1
    # via
    #   -r .\requirements.in
    #   openai
hypercorn==0.17.3
    # via -r .\requirements.in
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.5.0
    # via opentelemetry-api
iniconfig==2.0.0
    # via pytest
isodate==0.7.2
    # via
    #   azure-storage-blob
    #   openapi-core
itsdangerous==2.2.0
    # via -r .\requirements.in
jinja2==3.1.5
    # via
    #   -r .\requirements.in
    #   semantic-kernel
jiter==0.8.2
    # via openai
jsonschema==4.23.0
    # via
    #   openapi-core
    #   openapi-schema-validator
    #   openapi-spec-validator
jsonschema-path==0.3.4
    # via
    #   openapi-core
    #   openapi-spec-validator
jsonschema-specifications==2024.10.1
    # via
    #   jsonschema
    #   openapi-schema-validator
lazy-object-proxy==1.10.0
    # via openapi-spec-validator
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.6.0
    # via openapi-core
msal==1.31.1
    # via
    #   azure-identity
    #   fastapi-msal
    #   msal-extensions
msal-extensions==1.2.0
    # via azure-identity
multidict==6.1.0
    # via
    #   -r .\requirements.in
    #   aiohttp
    #   httpie
    #   yarl
mypy-extensions==1.0.0
    # via black
nest-asyncio==1.6.0
    # via semantic-kernel
numpy==2.2.3
    # via
    #   pandas
    #   scipy
    #   semantic-kernel
openai==1.63.0
    # via
    #   -r .\requirements.in
    #   semantic-kernel
openapi-core==0.19.4
    # via semantic-kernel
openapi-schema-validator==0.6.3
    # via
    #   openapi-core
    #   openapi-spec-validator
openapi-spec-validator==0.7.1
    # via openapi-core
opentelemetry-api==1.30.0
    # via
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   semantic-kernel
opentelemetry-sdk==1.30.0
    # via semantic-kernel
opentelemetry-semantic-conventions==0.51b0
    # via opentelemetry-sdk
packaging==24.2
    # via
    #   black
    #   deprecation
    #   prance
    #   pytest
pandas==2.2.3
    # via -r .\requirements.in
parse==1.20.2
    # via openapi-core
pathable==0.4.4
    # via jsonschema-path
pathspec==0.12.1
    # via black
platformdirs==4.3.6
    # via black
pluggy==1.5.0
    # via pytest
portalocker==2.10.1
    # via msal-extensions
prance==23.6.21.0
    # via semantic-kernel
priority==2.0.0
    # via hypercorn
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
psutil==7.0.0
    # via -r .\requirements.in
pybars4==0.9.13
    # via semantic-kernel
pycparser==2.22
    # via cffi
pydantic==2.10.6
    # via
    #   fastapi
    #   fastapi-msal
    #   openai
    #   pydantic-settings
    #   semantic-kernel
pydantic-core==2.27.2
    # via pydantic
pydantic-settings==2.7.1
    # via
    #   fastapi-msal
    #   semantic-kernel
pygments==2.19.1
    # via
    #   httpie
    #   rich
pyjwt[crypto]==2.10.1
    # via
    #   msal
    #   pyjwt
pymeta3==0.5.1
    # via pybars4
pysocks==1.7.1
    # via requests
pytest==8.3.4
    # via
    #   -r .\requirements.in
    #   pytest-asyncio
    #   pytest-cov
pytest-asyncio==0.25.3
    # via -r .\requirements.in
pytest-cov==6.0.0
    # via -r .\requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via
    #   -r .\requirements.in
    #   pydantic-settings
python-multipart==0.0.20
    # via -r .\requirements.in
pytz==2025.1
    # via pandas
pywin32==308
    # via portalocker
pyyaml==6.0.2
    # via jsonschema-path
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-path
    #   jsonschema-specifications
regex==2024.11.6
    # via tiktoken
requests[socks]==2.32.3
    # via
    #   -r .\requirements.in
    #   azure-core
    #   httpie
    #   jsonschema-path
    #   msal
    #   prance
    #   requests-toolbelt
    #   tiktoken
requests-toolbelt==1.0.0
    # via httpie
rfc3339-validator==0.1.4
    # via openapi-schema-validator
rich==13.9.4
    # via httpie
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
ruamel-yaml==0.18.10
    # via prance
ruamel-yaml-clib==0.2.12
    # via ruamel-yaml
scipy==1.15.1
    # via semantic-kernel
semantic-kernel==1.21.1
    # via -r .\requirements.in
six==1.17.0
    # via
    #   azure-core
    #   prance
    #   python-dateutil
    #   rfc3339-validator
sniffio==1.3.1
    # via
    #   anyio
    #   openai
soupsieve==2.6
    # via beautifulsoup4
starlette==0.45.3
    # via
    #   -r .\requirements.in
    #   fastapi
    #   fastapi-msal
tiktoken==0.9.0
    # via -r .\requirements.in
tqdm==4.67.1
    # via openai
typing-extensions==4.12.2
    # via
    #   anyio
    #   azure-core
    #   azure-cosmos
    #   azure-identity
    #   azure-storage-blob
    #   beautifulsoup4
    #   fastapi
    #   openai
    #   opentelemetry-sdk
    #   pydantic
    #   pydantic-core
    #   referencing
tzdata==2025.1
    # via
    #   faker
    #   pandas
urllib3==2.3.0
    # via requests
uvicorn==0.34.0
    # via -r .\requirements.in
werkzeug==3.1.3
    # via
    #   -r .\requirements.in
    #   openapi-core
wrapt==1.17.2
    # via deprecated
wsproto==1.2.0
    # via hypercorn
xmlformatter==0.2.8
    # via -r .\requirements.in
yarl==1.18.3
    # via aiohttp
zipp==3.21.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
